fix(microvm): close a credential race and two fail-open egress translations
From review. The three bugs are real; I confirmed each against the code before
fixing, and the two egress ones by running them.

**The preload script was world-readable for a window** (`write_preload`).
`fs::write` created it at the umask default, wrote every brokered secret into
it, and only then chmod'd to `0600`. Any local user who opened it in between
read the credentials — the exact exposure the module goes out of its way to
avoid by keeping them out of `msb run`'s argv, so the same threat model
applies. It is now created `0600` by `OpenOptions::mode` before a byte is
written, with a `remove_file` first so a leftover from a recycled pid cannot be
adopted with a wider mode. The filename gains the run sequence too, matching
`SandboxGuard::new`, since a pid alone repeats across invocations.

**An out-of-range port widened the rule instead of failing.**
`example.com:99999` split off an all-digit suffix, `parse::<u16>()` failed,
`.ok()` turned that into `None`, and the emitted token was `allow@example.com`
— *every* port, from an entry that asked for one. Confirmed by running it:

    example.com:99999 -> Ok(["allow@example.com"])
    example.com:65536 -> Ok(["allow@example.com"])

Now an error. The docstring already promised "an entry we cannot translate
exactly is an error, never a rule that quietly allows more"; the code did the
opposite. Profile `net_egress` never passes through `validate_egress_rule`
(that guards `box allow` only), so a `.h5i/env.toml` reached this directly.

**An IPv6 literal was mangled rather than refused.** `2001:db8::1` split into
host `2001:db8:` and port `1`, emitting `allow@2001:db8::tcp:1` — an address
nobody wrote. The comment claimed the all-digit check protected IPv6; it does
not. An unbracketed entry with more than one `:` now joins the `,`/`@`
rejection list.

Two nits from the same review, both fixed:

- `close_inherited_fds` passed `c_uint` into the variadic `libc::syscall`,
  which reads `long`. Correct today by ABI zero-extension; now correct by
  construction.
- The auth proxy's in-flight counter was decremented at the end of the worker
  closure, so any path that did not reach it leaked a slot permanently — and
  after `MAX_IN_FLIGHT` leaks the proxy answers 503 forever and the box loses
  authenticated egress. Worse than the reported `spawn`-failure case: a panic
  inside `handle_client` leaked one too. It is an RAII guard now, released on
  unwind.

And the two "worth a second look" items, which were about visibility rather
than behaviour:

- **The macOS browser profile's temp-dir grant is now in MANUAL.md**, not only
  in the source. It says what it costs — the directory is shared with host
  processes and other browser boxes, reintroducing for one profile on one
  platform the rendezvous point the `/tmp` redirect exists to remove. On the
  digest question the reviewer raised: `load_policy` digests the *stored*
  `policy.resolved.toml` against the digest written beside it at create time and
  never re-resolves, so a machine-specific path in it breaks nothing — but a
  `browser` env pulled to a second Mac carries a grant for a directory that does
  not exist there and Chrome will not start until it is recreated. That is now
  documented too. The same section drops the claim that the profile "has not
  been exercised on macOS at all", which is no longer true, and replaces it with
  what actually fails there and why.
- **The private-path filter is stated rather than inferred.** `box status` lists
  the paths (`private : cache (per-env; excluded from diff/export)`), and
  `box diff` prints a `note: N path(s) excluded as private (...)` line. That
  note goes to **stderr**, deliberately: `diff`'s return value is what `export`
  writes to `patch.diff`, so a footer inside it would be a footer inside a
  patch, and `h5i box diff > x.patch` still produces a file `git apply` accepts
  (verified).

Not taken: the `MAX_REQUEST_BODY x MAX_IN_FLIGHT` ~2 GiB ceiling. A shared byte
budget would be tighter, but it is a design change to a bounded-and-documented
limit rather than a defect, and it belongs in its own change.

Suite 573 passing (three new regression tests), clippy clean, all CI gates run
from a clean build.

+++ crates/h5i-sandbox/src/sandbox.rs
@@ -1932,7 +1932,17 @@ pub(crate) struct EgressJail {
 unsafe fn close_inherited_fds() {
     // close_range(2) is Linux 5.9+; Landlock needs 5.13+, so on any host that
     // reaches this code it is present and this is the only branch that runs.
-    if libc::syscall(libc::SYS_close_range, 3 as libc::c_uint, libc::c_uint::MAX, 0) == 0 {
+    // `syscall` is variadic, so each argument is read as a `long`. Passing a
+    // 32-bit `c_uint` happens to work on x86-64 and aarch64 because the ABI
+    // zero-extends, but the widths should agree by construction rather than by
+    // the platform being forgiving.
+    if libc::syscall(
+        libc::SYS_close_range,
+        3 as libc::c_long,
+        libc::c_uint::MAX as libc::c_long,
+        0 as libc::c_long,
+    ) == 0
+    {
         return;
     }
     // Fallback for a kernel (or seccomp policy) without close_range: walk the
